Small Re-Organize component
Divided frontend component page wise .
