+++ resources/js/components/frontend/home-page/production-process-section.tsx
