+++ resources/js/components/frontend/home-page/slider-area.tsx
