+++ resources/js/components/frontend/home-page/technologies-section.tsx
