+++ resources/js/pages/admin/customize-home/customize-home-slider.tsx
@@ -3,7 +3,7 @@ import AdminLayout from '@/layouts/admin/admin-layout';
 export default function Customizeslider() {
     return (
         <div>
-            <AdminLayout title="Dashboard">
+            <AdminLayout title="">
                 customize-Customizeslider
             </AdminLayout>
         </div>
